refactor: make ListItem fully reusable
