+++ packages/widget/src/pages/SendToWallet/SendToWalletPage.style.tsx
@@ -1,26 +1,19 @@
-import type { Theme, MenuProps } from '@mui/material';
+import type { Theme } from '@mui/material';
 import {
   InputBase,
   Box,
   Button,
   Typography,
   IconButton,
-  ButtonBase,
-  Menu,
-  menuClasses,
-  menuItemClasses,
-  svgIconClasses,
   inputBaseClasses,
   styled,
   List,
-  ListItem as MuiListItem,
-  containerClasses,
 } from '@mui/material';
 import { Card } from '../../components/Card';
 import type { PageContainerProps } from '../../components/PageContainer';
 import { PageContainer } from '../../components/PageContainer';
 
-export const NameInput = styled(InputBase)(({ theme }) => ({
+export const NameInput = styled(InputBase)(() => ({
   alignItems: 'start',
   [`.${inputBaseClasses.input}`]: {
     fontWeight: 500,
@@ -28,7 +21,7 @@ export const NameInput = styled(InputBase)(({ theme }) => ({
     paddingBottom: 0,
   },
 }));
-export const AddressInput = styled(NameInput)(({ theme }) => ({
+export const AddressInput = styled(NameInput)(() => ({
   minHeight: 64,
 }));
 
@@ -156,50 +149,6 @@ export const BookmarkButtonContainer = styled(Box)(({ theme }) => ({
   padding: theme.spacing(0, 1, 2),
   marginBottom: theme.spacing(-1.5),
 }));
-
-export const ListItemContainer = styled(MuiListItem)(() => ({
-  position: 'relative',
-  padding: 0,
-}));
-
-export const ListItemInfoContainer = styled(Box)(({ theme }) => ({
-  display: 'flex',
-  margin: 0,
-  justifyContent: 'flex-start',
-  alignItems: 'center',
-  gap: theme.spacing(1.5),
-}));
-export const ListItemMenuButton = styled(IconButton)(({ theme }) => ({
-  position: 'absolute',
-  right: theme.spacing(2),
-  '&:hover, &:focus': {
-    background:
-      theme.palette.mode === 'light'
-        ? theme.palette.grey[200]
-        : theme.palette.grey[800],
-  },
-}));
-
-export const ListMenu = styled(Menu)(({ theme }) => ({
-  [`& .${menuClasses.paper}`]: {
-    borderRadius: theme.shape.borderRadius,
-    boxShadow: '0px 2px 4px rgb(0 0 0 / 8%), 0px 8px 16px rgb(0 0 0 / 8%)',
-    padding: theme.spacing(2),
-    [`& .${menuClasses.list}`]: {
-      padding: 0,
-    },
-    [`& .${menuItemClasses.root}`]: {
-      borderRadius: theme.shape.borderRadiusSecondary,
-      padding: theme.spacing(1, 2, 1, 1),
-      [`& .${svgIconClasses.root}`]: {
-        fontSize: 20,
-        color: theme.palette.text.primary,
-        marginRight: theme.spacing(1.5),
-      },
-    },
-  },
-}));
-
 export const EmptyContainer = styled(Box)(({ theme }) => ({
   display: 'flex',
   flexDirection: 'column',
